functional map with location markers
