+++ src/app/app.component.html
@@ -5,7 +5,7 @@
 
 <div class="main-container">
   <main class="grid-container">
-    <div class="ro-container">
+    <div class="ro-container" fxFlexLayout="column" fxFlexAlign="center center">
       <router-outlet></router-outlet>
     </div>
     <footer>
