--- src/app/components/landing-page/google-map-rev.component.html
@@ -1,7 +1,0 @@
-
-<google-map *ngIf="apiLoaded |async"
-    height="400px"
-    width="750px"
->
-</google-map>
-
