+++ src/app/components/landing-page/google-map/google-map-rev.component.sass
@@ -1,0 +1,2 @@
+google-map div.map-container
+    width: 100%
