+++ src/app/components/landing-page/landing-page.component.html
@@ -1,12 +1,12 @@
-<div id="landingPage" fxLayout='column' fxLayoutAlign="center center">
-  <section>
-    <div class="title pb-2">
-      <h3>Welcome to AssignForce</h3>
-      <aside>Space Management Platform</aside> 
+<div id="landingPage" fxLayout='column' fxLayoutAlign="center center" fxLayoutGap="1vh">
+  <section id="landingWelcome" fxLayout='column' fxLayoutAlign="center center">
+    <div fxLayout fxLayoutAlign="center">
+      <h2>Welcome to AssignForce</h2>
+      <google-map-rev></google-map-rev>
     </div>
   </section>
-  <section>
-    <div class="title">
+  <section id="landingAbout" fxLayout fxLayoutAlign="center">
+    <div>
       <h3>About</h3>
       <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Cumque, possimus! Voluptatem, sint cumque maiores quis
         tenetur nam minima sapiente aliquid. Aspernatur porro nesciunt eius minus, rerum, officiis sunt labore neque hic
