+++ src/app/components/landing-page/landing-page.component.ts
@@ -4,6 +4,6 @@ import { Component } from '@angular/core';
   templateUrl: './landing-page.component.html',
   styleUrls: ['./landing-page.component.sass']
 })
-export class LandingPageComponent {
+export class LandingPageComponent { 
   constructor() { }
 }
